Add files via upload
Reset this to only predict for rank (1, 2, 3, 4, or 5 as one-hot encoded variables) using 'currently_insured', 'number_of_vehicles', 'number_of_drivers', and 'marital_status' as we talked about today. At the bottom is the ability to plot the accuracy, precision, and recall vs. the logistic regression probability cutoffs. The predictive power is pretty bad (for precision and recall anyway) for pretty much all ranks, just thought I'd include this for the sake of the plots. For rank 5 it is probably the best, if you choose the cutoff correctly.